Move modules directory to site-modules for clarity
This commit renames the 'modules' directory to 'site-modules'.  This is
the original directory name from the puppetlabs control-repo template.
The name was changed to 'modules' to match the puppeteers blog post.
However, 'site-modules' is more descriptive and less confusing when it
comes to differentiating it from the '.modules' directory created by the
`bolt module install` command.